Updated for a no embed option

+++ src/commands/Utility/dictionary.js
@@ -6,10 +6,16 @@ exports.run = function (bot, msg, args) {
         throw 'Please provide a word to search!';
     }
 
-    let word = args[0];
+    let parsed = bot.utils.parseArgs(args, ['e']);
+
+    let word = parsed.leftover[0];
 
     webdict('dictionary', word)
         .then(resp => {
+            if(parsed.options.e) {
+                msg.edit(resp.definition[0]);
+                return;
+            }
             msg.delete();
             msg.channel.sendEmbed(
                 bot.utils.embed(
